pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "openhands-server"
version = "0.1.0"
description = "OpenHands App Server"
readme = "README.md"
license = { text = "MIT" }
authors = [ { name = "OpenHands Team", email = "contact@all-hands.dev" } ]
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "aiosqlite>=0.19",
  "alembic>=1.13",
  "cloud-sql-python-connector>=1.18.4",
  "docker>=7.1,<8",
  "fastapi>=0.104",
  "google-api-python-client>=2.181",
  "google-auth-httplib2>=0.2",
  "google-auth-oauthlib>=1.2.2",
  "greenlet>=3.2.4",
  "openhands-agent-server @ git+https://github.com/All-Hands-AI/agent-sdk.git@6dd2af9e7f31e9b5488321d2eb1ae9e1f357e043#subdirectory=openhands/agent_server",
  "openhands-sdk @ git+https://github.com/All-Hands-AI/agent-sdk.git@6dd2af9e7f31e9b5488321d2eb1ae9e1f357e043#subdirectory=openhands/sdk",
  "openhands-tools @ git+https://github.com/All-Hands-AI/agent-sdk.git@6dd2af9e7f31e9b5488321d2eb1ae9e1f357e043#subdirectory=openhands/tools",
  "pydantic>=2",
  "sqlalchemy>=2",
  "uvicorn>=0.31.1",
  "websockets>=12",
]

optional-dependencies.dev = [
  "black>=23",
  "flake8>=6",
  "httpx>=0.25",
  "isort>=5",
  "mypy>=1",
  "pre-commit>=3.7",
  "pytest>=7",
  "pytest-asyncio>=0.21",
  "ruff>=0.11.8",
]
scripts.openhands-sdk-server = "openhands_server.agent_server.__main__:main"

[dependency-groups]
dev = [
  "httpx>=0.25",
  "openapi-python-client>=0.26.1",
  "pre-commit>=4.3",
  "pycodestyle>=2.12",
  "pyinstaller>=6.15",
  "pyright>=1.1.405",
  "pytest>=8.4.1",
  "pytest-asyncio>=0.21",
  "ruff>=0.12.10",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [ "openhands_server" ]

# Ruff configuration

[tool.ruff]
target-version = "py312"
line-length = 88

format.indent-style = "space"
format.quote-style = "double"
lint.select = [ "E", "F", "I" ]
lint.isort.combine-as-imports = true
lint.isort.force-single-line = false
lint.isort.known-first-party = [ "openhands", "openhands_server" ]
lint.isort.lines-after-imports = 2

[tool.uv.sources]
openhands-sdk = { git = "https://github.com/All-Hands-AI/agent-sdk.git", subdirectory = "openhands/sdk", rev = "6dd2af9e7f31e9b5488321d2eb1ae9e1f357e043" }
openhands-tools = { git = "https://github.com/All-Hands-AI/agent-sdk.git", subdirectory = "openhands/tools", rev = "6dd2af9e7f31e9b5488321d2eb1ae9e1f357e043" }
openhands-agent-server = { git = "https://github.com/All-Hands-AI/agent-sdk.git", subdirectory = "openhands/agent_server", rev = "6dd2af9e7f31e9b5488321d2eb1ae9e1f357e043" }
